a function that frees a listint_t list.

+++ 0x13-more_singly_linked_lists/5-free_listint2.c
@@ -11,6 +11,8 @@ void free_listint2(listint_t **head)
 {
 	listint_t *delete_all;
 
+	if (!head)
+		return;
 	while (*head)
 	{
 		delete_all = *head;
